Factored out code relating to getting info about network interfaces.  This commit includes the code itself, but no changes to the surrounding codebase to actually use it.  Makefile updates are included, so this new code is built.
I had to change Exception.h to make its
destructor virtual, so it is safe to subclass,
and changed its private members to protected,
so subclasses could access.  OutOfMemoryException
in particular needed access to the real
exception string, not a temporary, as the
GetErrorMessage() method was coded to work.

The src/unix/ subdirectory is intended to hold code
common to all unix-like OSs we support.
NetworkInterfaces.h is there, since I don't plan
to change any windows code to use it (yet).  The
implementations of the interface in that header
are in the platform-specific directories, not in
src/unix/.