Added support for passing through the applicability_check attribute on criteria, criterion, and extend_definition tags.
No deduction is attempted when the attribute
is missing, so the applicability check value
on Criteria, Criterion, and ExtendedDefinition
objects will not always be accurate, according
to proposed semantics for the attribute.  But
the interpreter doesn't use the attributes for
anything at this point, so an implementation
that has the effect of merely passing through
the attribute values seems to be good enough
for now.