Unix file finder consolidation.  All the unix FileFinder classes were identical, so I deleted them from the src/macos and src/solaris dirs, and moved the src/linux copy into src/unix.
Unfortunately, in order to allow the independent
probes to continue to work as-is, I had to
add the unix directory to the compile include path.
Otherwise, "#include <FileFinder.h>" would not
work.  I happen to like some hierarchical
organization in the source directories, but
you can't have "unix/FileFinder.h" in source
that is supposed to work on windows.  And I
didn't want to add more preprocessor junk...

It turned out to be kind of awkward to have a
file which is supposed to be transparently
swappable to a windows version, mixed in with
other files which are not.  Maybe the organization
of source code could be improved further, to
more cleanly separate them.  E.g. have
something like "swappable/unix" be separate
from "unix", and also have "swappable/windows",
etc.  Or maybe "platform/unix", "platform/windows",
etc...?  Anyway, no one probably cares too much
at this point, and I am shortly going to be off
the project for 3 months...