Bulk commit, mostly having to do with enabling querying of the 64-bit view of the registry in a 32-bit build of the interpreter.  Also:
- Changed the RegKey class, inlining most methods, enabling const-correctness,
and adding a ToString() method which essentially duplicates 
RegistryFinder::BuildRegistryKey().  I thought I should be able to get a string
representation of RegKey's without needing the help of some other class.

- Moved the WOW check for the purposes of printing a warning message when the
program starts up, from Main.cpp into WindowsCommon.  That code is a bit more
robust now, and checks whether the "IsWow64Process" function exists before
trying to call it.

- Added Is64BitOS() to WindowsCommon so windows_view values can be selected
based on the bitness of the OS.

- The regKeyObjectType stuff is removed from RegistryFinder; SE_REGISTRY_KEY
is used in all cases, since it seems to always work.  The ..._WOW64_32KEY type
only worked under certain circumstances, so it was simpler to just not use it
at all.