Fixed more bugs in the registry-related probes, mostly security-related.  It turns out GetNamedSecurityInfo() was automatically addressing certain security-related things, and since I moved away from using that function, I had to start doing it myself.
- Added an access parameter to the RegistryFinder::GetHKeyHandle()
methods, since some operations with the handle depend on what
access was specified at the time it was opened.  So I needed the
flexibility to be able to specify them at the time of opening.

- Added an isDirectory param to FileFinder::GetFileHandle(), since
getting handles to directories requires a slightly different
incantation compared to getting handles to files.  So the method
needed to know which incantation to use.  I could have 
auto-detected it with GetFileAttributes(), but it seems like at
the time of call, the caller always knew that.  So it was faster
and simpler to just pass it in.

- Added a new PrivilegeGuard class, to allow exception-safe
acquisition and release of privileges.

- Removed privilege escalation code from 
WindowsCommon::GetAuditedPermissionsForWindowsObject(), since
it no longer works to escalate the privileges at that point.
You have to do it earlier, before you get the handle you're calling
the GetSecurityInfo() function on.

- Added/replaced privilege escalation code to several windows
probes, in places where it was used exception-unsafely, or where
it was necessary to make the probe work.