Adding support for windows_view behavior in file-related probes.  64-bit interpreter builds can't query the 32-bit view yet.  The independent probes still need testing on non-windows platforms.  More details:
- Added the FsRedirectionGuard class to enable exception-safe
disabling and re-enabling of filesystem redirection.  Use of
this class is mainly via macros in the source, since the code
tends to be similar everywhere.
- AbsFileFinder.cpp: applied a fix from Dave Rothenberg that
fixes a partial matching bug.
- AbsFileFinder.cpp: removed some code I'd tentatively
commented out awhile ago... it had no effect, as I thought,
so now it's really gone.
- The BitnessView enumeration and some related code is moved
from RegistryFinder to WindowsCommon, since it is not specific
to the registry.  This caused minor changes to all registry-
related probes.
- Privilege escalation and release before/after the filefinder
runs is changed to use PrivilegeGuard in the file-related
probes.
- FileHash58Probe.cpp: there seems to have been an extraneous
call to Object::Analyze() in there.  From the old comments, I
think that was the remains of an earlier code design.  I removed
it.
- FileProbe: streamlined some error handling.  Files which no
longer exist after FileFinder indicated they did, should no
longer result in errors.  There's nothing wrong with that; the
filesystem isn't frozen while ovaldi is scanning it, so files
could pop in and out of existence at any time.
- FileProbe: added windows error messages to some errors.  I
dunno why it wasn't there before; maybe 
WindowsCommon::GetErrorMessage() wasn't written yet when the
file probe was developed?
- FileProbe: Corrected error handling for LookupAccountSid.
The return value of the first call wasn't being checked.  Also
a BOOL return value was being stored in a DWORD var.
- FileProbe: fixed a memory leak: the security descriptor from
GetNamedSecurityInfo() (now GetSecurityInfo()) was not being
freed.
- FileProbe: replaced snprintf() with a safer
Common::ToString().
- FileMd5Probe: not tested, since that test seems to have been
removed from the schema.  Definitions with that test won't
validate.  Should we remove the probe??