Fixed REGEX::RemoveExtraSlashes() to remove all escaping backslashes, not just those which were escaping another backslash.
Fixed windows FileFinder::FindPaths() to not throw an
exception if the constant portion of a path regex doesn't
exist.  Now just doesn't return any search results, as it
should have done.

Fixed a different bug in the same method where if a regex
was entirely constant with no pattern part, and the path
exists, it wasn't getting cased properly.  Now it is.