More printf removal.  Also:
windows fileprobe: removed a lot of copy-n-paste,
so in the future the same fixes don't need to be
duplicated in 6 different places...

textfilecontentprobe: removed the SNPRINTF macro.
It wasn't used anywhere.

windows ldapprobe: I didn't test this fix since we
don't seem to have any test content for it.  But I
pasted the code from Digest.cpp which I know works.
So I am pretty sure it's fine.