Changed the error message when a 32-bit ovaldi is trying to get the 32-bit processes.  The old error just said there was an error opening the process, but you can also get that error message in individual items, so it is unclear why sometimes that error results in aborting collection altogether and sometimes it doesn't.
The important difference is that in the first situation,
the error occurred while obtaining the list of 32-bit
processes to check.  If you can't do that, you know it is
impossible to get reliable results, so there's no point
in continuing.  On the other hand, if you know you
got all the processes, then you have a chance of getting
complete and reliable results, so it makes more sense to
continue collection.  You'll still get an error flag on
the collected object, but you will see which pids, from
the complete set, could and couldn't be opened.