Removed useless loops, now that values which aren't found are no longer included in partial match items.
Also, when a filepath is used and none are found,
I took out code that sets the filename entity to
other than ""/doesnotexist.  If filepath is not
found, filename can't be anything else.

Also, since other similar probes weren't adding
a full set of item entities to their partial match
items (i.e. those which correspond to object entities),
I decided to not do it in this probe either.  So
I removed a bunch of code for adding hash_type item
entities on partial match items.