Removed useless loops, now that values which aren't found are no longer included in partial match items.
This probe hadn't been fixed to use empty values when
status=does not exist, so I fixed it.

Also, when a filepath is used and none are found,
I took out code that sets the filename entity to
other than ""/doesnotexist.  If filepath is not
found, filename can't be anything else.