Simplified XslCommon.cpp, using a much simpler API (a C++ API too, not a C API).
XmlProcessor no longer manages Xerces initialization.  It
seems like whoever wrote that assumed that all xml processing
goes through that class, which is why it made sense for
initialization to go there.  In reality, XSLT processing done
by Xalan is done on XML and uses Xerces, so XML processing
is not the exclusive domain of XmlProcessor.  So it doesn't
make sense for it to go there.  If you delete XmlProcessor,
which shuts down Xerces, and then try to do an XSL transform,
you get a crash.  This is what ovaldi actually did.  I guess it
worked before, although I'm not sure how (maybe the C API 
re-init'd it again??), but it doesn't work now.  So now, Xerces
init is part of the standard setup steps of ovaldi, in Main.cpp.
It is done in an exception-safe way using an RAII class.