Improved windows account group expansion.  Now, the group is checked to try to determine whether it's global or local, before deciding which API calls to use to get its membership.  This is done based on SID structure. Also:
- To improve performance, some things are computed once at
app startup, and cached, so they can be reused.  This
includes computer and domain name, computer and domain SID,
and domain controller hostnames (if the computer is joined
to a domain).  There is code added to Main.cpp to init and
uninit this cached data, in WindowsCommon.  The local vs
global SID check reuses the cached computer SID.

- WindowsCommon::GetDomainControllerName() was ignoring errors
and just returning NULL, which causes callers to search the
local computer.  This can cause the wrong trustee to be found,
depending on the API used.  But always throwing or logging a
message on error meant that when the domain was the local 
computer, you'd see an error, since NetGetAnyDCName() errors
out on that "domain".  Of course local trustees which are
prefixed by the computer name must be handled properly.  This
means that you need to check whether the domain is the local
computer before you attempt to get a domain controller and
potentially error out.  In essence, the same local vs global
check needs to be made as described above for group expansion.

This could be done by converting to SIDs and using the same code
as above.  But so much is already done in terms of names, that
I just added some checks of trustee domains against the computer
name in various places... ideally as much as possible of this
code would work in terms of SIDs, but that's not how it is now.

- Added an overload of Common::EqualsIgnoreCase() which works on
wide strings.

- Added a wide-char version of WindowsCommon::GetSIDForTrusteeName()
since I had wide char strings and am tired of having to narrow them
all the time to work with narrow string APIs...

- Added a NetApiBufferFreeGuard class inside WindowsCommon to ensure
NetApiBufferFree() gets called when necessary... a lot of these
windows account functions require calling that to free returned
resources.  It's not used everywhere it could be; its usage could
be expanded... but not right now.