Re-adding windows services probe to the probefactory, since the merge seems to have removed that.
Fixes to restore the probe to compilability: this was a change to
variable handling, made in accordance with variable value memory
leak fixes which were merged from trunk.

Moved WindowsServicesProbe.h back to the "Header Files\probes"
project folder.  That seems to have been undone in the merge as
well.

Minor #include changes...